+++ src/pages/Message.jsx
@@ -1,142 +1,32 @@
-import { useParams } from "react-router";
-import { useState, useRef, useCallback, useEffect } from "react";
 import Container from "../components/Container/Container";
 import MessageInput from "../features/Message/MessageElements/MessageInput";
 import MessageProfile from "../features/Message/MessageElements/MessageProfile";
 import MessageSelect from "../features/Message/MessageElements/MessageSelect";
 import MessageEditor from "../features/Message/MessageElements/MessageEditor";
 import Button from "../components/Button/Button";
-import useMessageProfile from "../features/Message/hooks/useMessageProfile";
-import useMessageSubmit from "../features/Message/hooks/useMessageSubmit";
+import useMessage from "../features/Message/hooks/useMessage";
 
 const style = {
   font: "text-24 font-bold text-gray-900 mb-[15px]",
 };
 
 const Message = () => {
-  const { id } = useParams();
-  const editorRef = useRef(null);
-  const [errorMsg, setErrorMsg] = useState("");
-  const [selectedProfile, setSelectedProfile] = useState("");
-  const [profileOptions, setProfileOptions] = useState([]);
-  const [isDisable, setIsDisable] = useState(true);
-  const [inputSender, setInputSender] = useState("");
-  const [inputText, setInputText] = useState("");
-  const { images } = useMessageProfile();
-
-  const [postMessageData, setPostMessageData] = useState({
-    team: "18-4",
-    recipientId: id,
-    sender: "",
-    profileImageURL: "",
-    relationship: "지인",
-    content: "",
-    font: "Noto Sans",
-  });
-
-  useEffect(() => {
-    if (images.imageUrls) {
-      setPostMessageData((prev) => ({
-        ...prev,
-        profileImageURL: images.imageUrls[0],
-      }));
-      setProfileOptions(images.imageUrls.slice(1));
-    }
-  }, [images]);
-
-  const handleInputChange = useCallback((e) => {
-    const inputValue = e.target.value;
-    setPostMessageData((prev) => ({
-      ...prev,
-      sender: inputValue,
-    }));
-    if (errorMsg) {
-      setErrorMsg("");
-    }
-  }, []);
-
-  const handleInputBlur = (e) => {
-    const inputValue = e.target.value.trim();
-    const errorMsg = "값을 입력해 주세요.";
-
-    if (!inputValue) {
-      setErrorMsg(errorMsg);
-    } else {
-      setErrorMsg("");
-    }
-  };
-
-  const handleProfile = (imageUrl) => {
-    setSelectedProfile(imageUrl);
-    setPostMessageData((prev) => ({
-      ...prev,
-      profileImageURL: imageUrl,
-    }));
-  };
-
-  const handleSelectChange = (selectedOption) => {
-    setPostMessageData((prev) => ({
-      ...prev,
-      relationship: selectedOption.value,
-    }));
-  };
-
-  const handleTextChange = useCallback(() => {
-    if (editorRef.current) {
-      const html = editorRef.current.root.innerHTML;
-
-      setPostMessageData((prev) => ({
-        ...prev,
-        content: html,
-      }));
-    }
-  }, []);
-
-  const handleSelectionChange = useCallback((range, oldRange, source) => {
-    if (range && editorRef.current) {
-      const format = editorRef.current.getFormat();
-      let fontName = format.font;
-
-      if (
-        !fontName ||
-        Array.isArray(fontName) ||
-        typeof fontName !== "string"
-      ) {
-        fontName = "Noto Sans";
-      } else {
-        fontName = fontName.replace(/^['"]+|['"]+$/g, "");
-
-        if (!fontName) fontName = "Noto Sans";
-      }
-
-      setPostMessageData((prev) => ({
-        ...prev,
-        font: "Noto Sans",
-      }));
-    }
-  }, []);
-
-  const handleButtonDisable = (type = "", e) => {
-    if (type === "input") {
-      const input = e.target.value.trim();
-      setInputSender(input);
-      if (input && inputText) {
-        setIsDisable(false);
-      } else {
-        setIsDisable(true);
-      }
-    } else if (type === "" && editorRef.current) {
-      const text = editorRef.current.getText().trim();
-      setInputText(text);
-      if (inputSender && text) {
-        setIsDisable(false);
-      } else {
-        setIsDisable(true);
-      }
-    }
-  };
-
-  const { handleSubmit } = useMessageSubmit(postMessageData);
+  const {
+    editorRef,
+    errorMsg,
+    selectedProfile,
+    profileOptions,
+    isDisable,
+    postMessageData,
+    handleInputChange,
+    handleInputBlur,
+    handleProfile,
+    handleSelectChange,
+    handleEditorChange,
+    handleFontChange,
+    handleButtonDisable,
+    handleSubmit,
+  } = useMessage();
 
   return (
     <Container
@@ -145,7 +35,7 @@ const Message = () => {
     >
       <MessageInput
         style={style}
-        value={setPostMessageData.sender}
+        value={postMessageData.sender}
         onChange={handleInputChange}
         onBlur={(e) => {
           handleInputBlur(e);
@@ -169,8 +59,8 @@ const Message = () => {
         style={style}
         ref={editorRef}
         value={postMessageData.content}
-        onChange={handleTextChange}
-        onSelectionChange={handleSelectionChange}
+        onChange={handleEditorChange}
+        onSelectionChange={handleFontChange}
         onBlur={handleButtonDisable}
       />
       <Button
